Move pending loader, filter proposals, hooks for proposal tabs
