+++ components/proposals/form/ProposalForm.tsx
@@ -30,7 +30,7 @@ import UrlLink from "@/components/utils/UrlLink";
 import RequiredMark from "@/components/utils/RequiredMark";
 import Subtitle from "../subtitle/Subtitle";
 import LoadingCover from "@/components/utils/LoadingCover";
-import useProposals from "../hooks/useProposals";
+import useProposals from "./hooks/useProposals";
 import { twMerge } from "tailwind-merge";
 
 const ProposalForm = () => {
