+++ components/proposals/list/ListHome.tsx
@@ -1,16 +1,18 @@
 import CreatedDateString from "@/components/utils/CreatedDateString";
 import { TypeProposal } from "@/types/type";
-import React from 'react'
+import React from "react";
 import Tabs from "./Tabs";
-import useProposalTab from "./useProposalTab";
+import useProposalTab from "./hooks/useProposalTab";
 import { useSearchParams } from "next/navigation";
 import List from "./ProposalList";
+import LoadingSpinnerM from "@/components/utils/LoadingSpinnerM";
 
 type Props = {
-  proposals: TypeProposal[],
+  proposals: TypeProposal[];
+  isPending: boolean;
 };
 
-const ListHome = ({ proposals }: Props) => {
+const ListHome = ({ proposals, isPending }: Props) => {
   const { tabs, handleTabClick } = useProposalTab();
   const searchParams = useSearchParams();
   const currTab = searchParams.get("tab");
@@ -54,7 +56,7 @@ const ListHome = ({ proposals }: Props) => {
   return (
     <div className="flex flex-col gap-3">
       <Tabs tabs={tabs} handleTabClick={handleTabClick} />
-      {displayList()}
+      {isPending ? <LoadingSpinnerM /> : displayList()}
     </div>
   );
 };
