+++ components/proposals/list/NoData.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const NoData = () => {
+  return (
+    <div className="text-center">Nothing found</div>
+  )
+}
+
+export default NoData
