+++ components/proposals/list/ProposalList.tsx
@@ -1,11 +1,14 @@
 import CreatedDateString from "@/components/utils/CreatedDateString";
+import NoFetchedData from "@/components/utils/NoFetchedData";
 import { TypeProposal } from "@/types/type"
 import React from 'react'
+import NoData from "./NoData";
 
 type Props = {
   proposals: TypeProposal[];
 }
 const ProposalList = ({ proposals }: Props) => {
+  if (proposals.length === 0) return <NoData />;
   return (
     <div>
       {proposals.map((proposal) => (
